requirements.txt
pyzbar
Pillow
qrcode